+++ libs/react/src/lib/react.tsx
@@ -5,15 +5,8 @@ import React, {
   useContext,
   Context,
 } from 'react';
-import { debug } from 'debug';
-import { finalize, filter, tap } from 'rxjs/operators';
-import { Observable, of, concat } from 'rxjs';
-import {
-  Effect,
-  createSinks,
-  createSources,
-  SpawnEffect,
-} from '@rx-store/core';
+import { Observable } from 'rxjs';
+import { Effect, spawnRootEffect } from '@rx-store/core';
 
 /**
  * A React hook that consumes from the passed Rx Store context,
@@ -23,94 +16,6 @@ export const useStore = <T extends {}>(context: Context<T>): T => {
   const value = useContext(context);
   if (!value) throw new Error();
   return value;
-};
-
-// Creates an observable that will be subscribed to *before*
-// the underlying effect is subscribed to, will immediately
-// run a devtools hook & complete.
-const before$ = (ref: { debugKey: string }) =>
-  of(null).pipe(
-    tap(() => {
-      // TODO - add a devtools hook here
-      debug(`rx-store:${ref.debugKey}`)('spawn');
-    }),
-    filter(() => false)
-  );
-
-// Creates an observable that will be subscribed to *after*
-// the underlying effect is subscribed to, will immediately
-// run a devtools hook & complete.
-const after = (ref: { debugKey: string }) => {
-  debug(`rx-store:${ref.debugKey}`)('teardown');
-};
-
-const ids: Record<string, number> = {};
-
-/**
- * The spawnRootEffect runs the root effect which means the effectFn is called
- * with curried sources, sinks, and its own spawnEffect function used
- * to spawn child effects recursively.
- *
- * Running effects does not involve subscribing to the observables they return,
- * which is the responsibility of the <Manager /> component to subscribe/unsubscribe.
- * Running the effects only involves calling the effectFn, passing in curried sources, sinks
- * and spawnEffect function.
- *
- * @param storeValue
- * @param rootEffectFn
- */
-export const spawnRootEffect = <T extends {}>(
-  storeValue: T,
-  rootEffectFn: Effect<T>
-) => {
-  /**
-   * spawnEffect closes over the `storeValue`. It takes in a `debugKey`
-   * and an effectFn.
-   *
-   * It curries sources, sinks, and a spawnEffect specific to the effectFn
-   * being spawned, such that we retain the stack or context of where each
-   * effect was spawned, all the way back to the root effect, for any effect,
-   * like a stack trace.
-   *
-   * The effect (observable) is also wrapped in before and after "hooks" for
-   * devtools, so devtools can known when the effect is actuall subscribed & torn down.
-   *
-   * The curried sources and sinks also have devtools hooks embedded in them such
-   * that devtools knows when each effect receives value(s), which subject(s) they came
-   * from, and which subject(s) each effect sinks data back into.
-   */
-  const spawnEffect: SpawnEffect<T> = (effect, { name }) => {
-    // Curries the sources and sinks with the debug key, to track this
-    // effects "inputs" and "outputs" in the devtools.
-    const sources = createSources(name, storeValue);
-    const sinks = createSinks(name, storeValue);
-
-    // Keeps the "context" intact, by appending to debugKey to create a path
-    // each time an effect creates a child effect by running it's curried `spawnEffect()`
-    const childSpawnEffect: SpawnEffect<T> = (effect, { name: childName }) => {
-      const curriedName = name + ':' + childName;
-      if (undefined === ids[curriedName]) {
-        ids[curriedName] = 1;
-      } else {
-        ids[curriedName]++;
-      }
-      const id = ids[curriedName];
-      return spawnEffect(effect, {
-        name: `${curriedName}${id === 1 ? '' : id}`,
-      });
-    };
-
-    // Run the effect function passing in the curried sources, sinks, and
-    // spawnEffect function for the effectFn to run any of its children effectFn
-    const effect$ = effect({ sources, sinks, spawnEffect: childSpawnEffect });
-
-    // Sandwich the effect between before and after streams, allowing devools
-    // hooks to run when the effect is subscribed & torn down.
-    const ref = { debugKey: name };
-    return concat(before$(ref), effect$).pipe(finalize(() => after(ref)));
-  };
-
-  return spawnEffect(rootEffectFn, { name: 'root' });
 };
 
 /**
